requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
msal

pandas
openpyxl
python-docx
pdfplumber
pdf2image
pytesseract
Pillow
beautifulsoup4
lxml
